# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "S1 APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get category list",
                "description": "Get category list",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Filter by category name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category list",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "category": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create new category",
                "description": "Create new category",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category was stored successfully.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "category": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing required field"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/posts": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create new post",
                "description": "Create new post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Provide in header request: Authorization: Bearer ACCESS_TOKEN",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "category_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "author",
                        "in": "formData",
                        "description": "author",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "content",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post was stored successfully.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "post": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing required field"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "Category": {
            "required": [
                "label",
                "name"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "Post": {
            "required": [
                "title",
                "category_id",
                "creator_id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "title": {
                    "description": "title",
                    "type": "string"
                },
                "category_id": {
                    "description": "category_id",
                    "type": "string"
                },
                "author": {
                    "description": "author",
                    "type": "string"
                },
                "content": {
                    "description": "content",
                    "type": "string"
                },
                "creator_id": {
                    "description": "creator_id",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}